Avoid re-extracting DataPtrs when forwarding values between Futures
In a bunch of cases we need to "forward" between one future and another, typically because we need to convert the type of the data (e.g., from Message to PyObject). In most of these cases the DataPtrs of the value don't change, and yet the new future must re-extract them from scratch. By allowing the user to obtain the vector of extracted DataPtrs from the old future, we can allow them to "shortcut" this step.

Differential Revision: [D28118298](https://our.internmc.facebook.com/intern/diff/D28118298/)

ghstack-source-id: 127929533
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57433