Reland: [TensorExpr] Add methods for inspecting generated code in `TensorExprKernel`.
The new methods allow to peak into bufferArgs which describe parameters
that codegen expects. This description includes info whether a given
parameter is a scalar var or a buffer and in case it's a buffer allows
to get the corresponding `Buf*` pointer from which we could get the
expected sizes.

Relanding #57074  which was reverted because I forgot to guard a new
test with `ifdef LLVM`.

ghstack-source-id: c8cecb57bf110eabb4880e5e363a1b2522ea5bf2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57560