[quant][graphmode][fx] Quantize the output of statically quantized fp16 op in QuantizeHandler
Summary:
Previously the output of statically quantized fp16 operator is not quantized in QuantizeHandler, which is not consistent with
the behavior of static int8 operators. Also it does not work well with reference functions, this PR
changes the fp16 static QuantizeHandler to quantize (call to(torch.float16)) in the QuantizeHandler, this also
makes the future support for reference functions easier.

Test Plan:
python test/test_quantization.py TestQuantizeFx
python test/test_quantization.py TestQuantizeFxOps

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 6f21e34dc9c09530830fa1bd161156ca2b9301ef
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58445