[pkg] Catch exceptions where dependency resolution gets invalid imports
Users can create invalid imports, like:
```
HG: in a top-level package
if False:
  from .. import foo
```

Since this code is never executed, it will not cause the module to fail to
load. But our dependency analysis walks every `import` statement in the AST,
and will attempt to resolve the (incorrectly formed) import, throwing an exception.

For posterity, the code that triggered this: https://git.io/JsCgM

Differential Revision: [D28543980](https://our.internmc.facebook.com/intern/diff/D28543980/)

ghstack-source-id: 129366860
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58573