Add tools script to list `torch` APIs
This adds a script that will run through everything in `dir(torch)` recursively and print out each path. We can run this for 1.8.1 and then again for 1.9.0 to determine the API difference between the two in order to ensure everything is properly documented and there are no secret BC-breaking changes.

ghstack-source-id: a1b4dc48de9c7d97cfb58e8c82e290ef12e435cb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58610